Works, for just one Slicer instance
To test the setup.
The missing piece is to find a way to inject a path to the Javascript when the container is created, in the right place (the one million dollar question). This path would be used in the "location" line for websockets section for the new container in the nginx.conf file.